+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Auth: Map login errors from backend error codes.
 - Deps: update `@angular/cli` to 20.3.19.
 - Deps: update `marked` to 17.0.4.
+- Deps (dev): update `ng-extract-i18n-merge` to 3.3.0. In `angular.json`, remove `projects.app.architect.extract-i18n.configurations`.
 
 ### Fixed